add module for ObjectMapper
