+++ avs-api/src/test/java/org/example/age/avs/api/test/StubAvsServiceModule.java
@@ -1,0 +1,24 @@
+package org.example.age.avs.api.test;
+
+import dagger.Binds;
+import dagger.Module;
+import io.undertow.server.HttpHandler;
+import org.example.age.avs.api.AvsApi;
+import org.example.age.avs.api.AvsApiModule;
+import org.example.age.common.service.data.DataMapperModule;
+import org.example.age.common.service.data.UserAgentAuthMatchDataExtractorModule;
+import org.example.age.test.service.data.TestAccountIdExtractorModule;
+
+/** Dagger module that binds dependencies for <code>@Named("api") {@link HttpHandler}</code>. */
+@Module(
+        includes = {
+            AvsApiModule.class,
+            TestAccountIdExtractorModule.class,
+            UserAgentAuthMatchDataExtractorModule.class,
+            DataMapperModule.class,
+        })
+public interface StubAvsServiceModule {
+
+    @Binds
+    AvsApi bindAvsApi(StubAvsService service);
+}
