+++ site-api/src/test/java/org/example/age/site/api/SiteApiTest.java
@@ -3,20 +3,14 @@ package org.example.age.site.api;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import dagger.Binds;
 import dagger.BindsInstance;
 import dagger.Component;
-import dagger.Module;
-import dagger.Provides;
 import io.undertow.Undertow;
-import io.undertow.server.HttpHandler;
 import java.io.IOException;
 import java.util.Map;
 import javax.inject.Named;
 import javax.inject.Singleton;
 import okhttp3.Response;
-import org.example.age.common.service.data.UserAgentAuthMatchDataExtractorModule;
 import org.example.age.data.certificate.AgeCertificate;
 import org.example.age.data.certificate.SignedAgeCertificate;
 import org.example.age.data.certificate.VerificationRequest;
@@ -26,10 +20,8 @@ import org.example.age.data.crypto.BytesValue;
 import org.example.age.data.crypto.DigitalSignature;
 import org.example.age.data.crypto.SecureId;
 import org.example.age.data.user.VerifiedUser;
-import org.example.age.data.utils.DataMapper;
-import org.example.age.site.api.test.StubSiteService;
+import org.example.age.site.api.test.StubSiteServiceModule;
 import org.example.age.test.server.undertow.TestUndertowModule;
-import org.example.age.test.service.data.TestAccountIdExtractorModule;
 import org.example.age.testing.client.TestClient;
 import org.example.age.testing.server.TestUndertowServer;
 import org.junit.jupiter.api.Test;
@@ -86,27 +78,8 @@ public final class SiteApiTest {
         return SignedAgeCertificate.of(certificate, signature);
     }
 
-    /** Dagger module that binds dependencies needed to create a <code>@Named("api") {@link HttpHandler}</code>. */
-    @Module(
-            includes = {
-                SiteApiModule.class,
-                UserAgentAuthMatchDataExtractorModule.class,
-                TestAccountIdExtractorModule.class
-            })
-    interface TestModule {
-
-        @Binds
-        SiteApi bindAvsApi(StubSiteService impl);
-
-        @Provides
-        @Singleton
-        static ObjectMapper provideObjectMapper() {
-            return DataMapper.get();
-        }
-    }
-
     /** Dagger component that provides an {@link Undertow} server. */
-    @Component(modules = {TestUndertowModule.class, TestModule.class})
+    @Component(modules = {TestUndertowModule.class, StubSiteServiceModule.class})
     @Singleton
     interface TestComponent {
